download files, rearrange directory structure
