+++ backend/climate/cmip6/scenariomip/env.sample
@@ -1,0 +1,6 @@
+S3_BUCKET=uw-climaterisklab
+S3_BASE_PREFIX=climate/raw_data/CMIP6/ScenarioMIP/
+VARIABLE=burntFractionAll
+SSP=126
+FILE_FORMAT=h5netcdf
+CRS=4326
