--- backend/climate/env.sample
@@ -1,5 +1,0 @@
-VARIABLE=burntFractionAll
-S3_BUCKET=uw-climaterisklab
-S3_PREFIX=climate/raw_data/cmip6/burntFractionAll/
-FILE_FORMAT=netCDF
-CRS=4326
