--- backend/climate/process_climate.py
@@ -1,16 +1,0 @@
-import xarray as xr
-import boto3
-
-
-def download_climate_data():
-    pass
-
-
-def main(s3_uris: str, file_format: str, crs: str) -> xr.Dataset:
-
-    
-    pass
-
-
-if __name__ == "__main__":
-    main()
